+++ core/src/test/resources/openapi/spec_with_mandatory_array_query_parameter.yaml
@@ -1,0 +1,24 @@
+openapi: 3.0.0
+info:
+  title: Sample API
+  version: 0.1.9
+paths:
+  /products:
+    get:
+      summary: get products
+      description: Get multiple products filtered by Brand Ids
+      parameters:
+        - name: brand_ids
+          in: query
+          required: true
+          schema:
+            items:
+              type: number
+            type: array
+      responses:
+        '200':
+          description: OK
+          content:
+            application/json:
+              schema:
+                type: string
